# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API For Permasalahan \u0026 Isu Strategis Services",
        "title": "Permasalahan \u0026 Isu Strategis Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "${HOST}:${PORT}",
    "basePath": "/",
    "paths": {
        "/isu_strategis": {
            "post": {
                "description": "Create a new isu strategis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Isu Strategis Service"
                ],
                "summary": "Create Isu Strategis",
                "parameters": [
                    {
                        "description": "Create Isu Strategis",
                        "name": "isu_strategis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.IsuStrategisCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/isu_strategis/{id}": {
            "get": {
                "description": "FindById a isu strategis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Isu Strategis Service"
                ],
                "summary": "FindById Isu Strategis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isu Strategis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a isu strategis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Isu Strategis Service"
                ],
                "summary": "Update Isu Strategis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Isu Strategis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Isu Strategis",
                        "name": "isu_strategis",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.IsuStrategisUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
        "/isu_strategis/{kode_opd}/{tahun_awal}/{tahun_akhir}": {
            "get": {
                "description": "FindAll a isu strategis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Isu Strategis Service"
                ],
                "summary": "FindAll Isu Strategis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kode OPD",
                        "name": "kode_opd",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tahun Awal",
                        "name": "tahun_awal",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tahun Akhir",
                        "name": "tahun_akhir",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.IsuStrategisResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "web.DataDukungRequest": {
            "type": "object",
            "required": [
                "data_dukung",
                "narasi_data_dukung"
            ],
            "properties": {
                "data_dukung": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.JumlahDataRequest"
                    }
                },
                "narasi_data_dukung": {
                    "type": "string"
                },
                "permasalahan_opd_id": {
                    "type": "integer"
                }
            }
        },
        "web.DataDukungResponse": {
            "type": "object",
            "properties": {
                "data_dukung": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.JumlahDataResponse"
                    }
                },
                "narasi_data_dukung": {
                    "type": "string"
                },
                "permasalahan_opd_id": {
                    "type": "integer"
                }
            }
        },
        "web.DataDukungUpdateRequest": {
            "type": "object",
            "properties": {
                "data_dukung": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.JumlahDataUpdateRequest"
                    }
                },
                "narasi_data_dukung": {
                    "type": "string"
                }
            }
        },
        "web.IsuStrategisCreateRequest": {
            "description": "Request Isu Strategis Create",
            "type": "object",
            "required": [
                "isu_strategis",
                "kode_bidang_urusan",
                "kode_opd",
                "nama_bidang_urusan",
                "nama_opd",
                "tahun_akhir",
                "tahun_awal"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isu_strategis": {
                    "type": "string"
                },
                "kode_bidang_urusan": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_bidang_urusan": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                },
                "permasalahan_opd": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.PermasalahanOpdRequest"
                    }
                },
                "tahun_akhir": {
                    "type": "string"
                },
                "tahun_awal": {
                    "type": "string"
                }
            }
        },
        "web.IsuStrategisResponse": {
            "description": "Response Isu Strategis",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isu_strategis": {
                    "type": "string"
                },
                "kode_bidang_urusan": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_bidang_urusan": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                },
                "permasalahan_opd": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.PermasalahanResponse"
                    }
                },
                "tahun_akhir": {
                    "type": "string"
                },
                "tahun_awal": {
                    "type": "string"
                }
            }
        },
        "web.IsuStrategisUpdateRequest": {
            "description": "Request Isu Strategis Update",
            "type": "object",
            "required": [
                "id",
                "isu_strategis",
                "kode_bidang_urusan",
                "kode_opd",
                "nama_bidang_urusan",
                "nama_opd",
                "tahun_akhir",
                "tahun_awal"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isu_strategis": {
                    "type": "string"
                },
                "kode_bidang_urusan": {
                    "type": "string"
                },
                "kode_opd": {
                    "type": "string"
                },
                "nama_bidang_urusan": {
                    "type": "string"
                },
                "nama_opd": {
                    "type": "string"
                },
                "permasalahan_opd": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.PermasalahanIsuStrategisUpdateRequest"
                    }
                },
                "tahun_akhir": {
                    "type": "string"
                },
                "tahun_awal": {
                    "type": "string"
                }
            }
        },
        "web.JumlahDataRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_data_dukung": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "number"
                },
                "satuan": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.JumlahDataResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "id_data_dukung": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "number"
                },
                "satuan": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
        "web.JumlahDataUpdateRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "jumlah_data": {
                    "type": "number"
                },
                "satuan": {
                    "type": "string"
                },
                "tahun": {
                    "type": "string"
                }
            }
        },
...
        "web.PermasalahanIsuStrategisUpdateRequest": {
            "type": "object",
            "properties": {
                "data_dukung": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.DataDukungUpdateRequest"
                    }
                },
                "permasalahan_opd_id": {
                    "type": "integer"
                }
            }
        },
        "web.PermasalahanOpdRequest": {
            "type": "object",
            "properties": {
                "data_dukung": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.DataDukungRequest"
                    }
                },
                "id_permasalahan": {
                    "type": "integer"
                }
            }
        },
        "web.PermasalahanResponse": {
            "type": "object",
            "properties": {
                "data_dukung": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.DataDukungResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "jenis_masalah": {
                    "type": "string"
                },
                "level_pohon": {
                    "type": "integer"
                },
                "masalah": {
                    "type": "string"
                }
            }
        },
...
        "web.WebResponse": {
            "description": "Web Response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}